Add support for reloading cert from file & externally renewed secret
By providing the --external-tls-renewal flag in combination with --in-cluster, the TLS cert is fetched from a secret but no k8s CSR api is invoked
for initial request or renewal. Useful for running in i.e. OpenShift which can provide auto-renewal of service certificates.

Also replaced existing file based load with this watcher implementation to automatically reload, in case the cert is volume mounted