Add option to pass configmap with mapping between SA and IAM role (#142)
This adds an option for the webhook to watch a configmap for additional service account to Role mappings. This is useful where there are tooling that creates IAM Roles and already know what service account should use them. In particular kOps already has this mapping. Adding the annotation to the service accounts is then just additional manual work.

* Add option to use a configmap as source of mappings between service account and IAM role.
* Creates a separate informer and cache for the configmap.
* Enforces precedence between service account annotations and configmap where service accounts have the higher precedence.